+++ common/src/main/java/com/niuktok/backend/common/entity/UserVideoView.java
@@ -14,13 +14,13 @@ import lombok.ToString;
 public class UserVideoView implements Serializable {
     @Id
     @GeneratedValue(generator = "JDBC")
-    private Integer id;
+    private Long id;
 
     @Column(name = "user_id")
-    private Integer userId;
+    private Long userId;
 
     @Column(name = "video_id")
-    private Integer videoId;
+    private Long videoId;
 
     private Integer count;
